requirements.txt
joblib==1.1.0
numpy==1.22.3
pandas==1.4.3
Pillow==9.3.0
scikit_learn==1.1.3
scipy==1.7.3
streamlit==1.13.0
